Move the sample function and samples_analaysis property to the base class
