--- tests/output/project_settings.json
@@ -1,37 +1,0 @@
-{
-    "header": {
-        "schema_version": "0.1.0",
-        "aces_transform_id": "",
-        "aces_user_name": "",
-        "camera_make": "",
-        "camera_model": "",
-        "iso": 800,
-        "temperature": 5600,
-        "additional_camera_settings": "",
-        "lighting_setup_description": "",
-        "debayering_platform": "",
-        "debayering_settings": "",
-        "encoding_colourspace": "",
-        "rgb_display_colourspace": "sRGB",
-        "cat": "CAT02",
-        "optimisation_space": "Oklab",
-        "illuminant_interpolator": "Linear",
-        "decoding_method": "Median",
-        "ev_range": [
-            -1.0,
-            0.0,
-            1.0
-        ],
-        "grey_card_reference": [
-            0.18,
-            0.18,
-            0.18
-        ],
-        "lut_size": 1024,
-        "lut_smoothing": 32
-    },
-    "data": {
-        "colour_checker": {},
-        "grey_card": {}
-    }
-}
